requirements.txt
# Core dependencies
click>=8.0.0
anthropic>=0.21.0
requests>=2.31.0

# Video processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0

# Utilities
python-dotenv>=1.0.0
pathlib2>=2.3.7; python_version < '3.4'

# Optional but recommended
tqdm>=4.66.0        # Progress bars
colorama>=0.4.6     # Colored output
rich>=13.0.0        # Better CLI formatting

# Development dependencies (optional)
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0